services/flask-app/app.py
from flask import Flask, jsonify
import os
import redis
import psycopg2
from datetime import datetime

app = Flask(__name__)

# Configuration
DATABASE_URL = os.getenv('DATABASE_URL')
REDIS_URL = os.getenv('REDIS_URL')

@app.route('/')
def home():
    return jsonify({
        'status': 'running',
        'service': 'Flask Application',
        'timestamp': datetime.utcnow().isoformat()
    })

@app.route('/health')
def health():
    """Health check endpoint for monitoring"""
    checks = {
        'flask': 'healthy',
        'database': check_database(),
        'redis': check_redis()
    }
    
    all_healthy = all(status == 'healthy' for status in checks.values())
    status_code = 200 if all_healthy else 503
    
    return jsonify({
        'status': 'healthy' if all_healthy else 'degraded',
        'checks': checks,
        'timestamp': datetime.utcnow().isoformat()
    }), status_code

def check_database():
    """Check PostgreSQL connection"""
    try:
        # TODO: Implement actual database connection check
        return 'healthy'
    except Exception as e:
        return f'unhealthy: {str(e)}'

def check_redis():
    """Check Redis connection"""
    try:
        # TODO: Implement actual Redis connection check
        return 'healthy'
    except Exception as e:
        return f'unhealthy: {str(e)}'

@app.route('/metrics')
def metrics():
    """Prometheus metrics endpoint"""
    # TODO: Implement actual metrics collection
    return """
# HELP flask_app_requests_total Total number of requests
# TYPE flask_app_requests_total counter
flask_app_requests_total 0

# HELP flask_app_up Application is running
# TYPE flask_app_up gauge
flask_app_up 1
"""

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5000, debug=os.getenv('FLASK_DEBUG', 'False') == 'True')
