services/flask-app/requirements.txt
Flask==3.0.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
prometheus-client==0.19.0
gunicorn==21.2.0
